+++ Graphs/gain_bucket_entry.h
@@ -8,14 +8,14 @@
 
 class GainBucketEntry {
  public:
-  GainBucketEntry() : gain(0), id(0), current_weight_vector_index(0) {}
-  GainBucketEntry(int gain, int id, const std::vector<int>& weight_vector,
+  GainBucketEntry() : gain(0.0), id(0), current_weight_vector_index(0) {}
+  GainBucketEntry(double gain, int id, const std::vector<int>& weight_vector,
                   int current_weight_vector_index = 0)
     : gain(gain), id(id),
       current_weight_vector_index(current_weight_vector_index) {
     all_weight_vectors.push_back(weight_vector);     
   }
-  GainBucketEntry(int gain, Node* node) : gain(gain), id(node->id),
+  GainBucketEntry(double gain, Node* node) : gain(gain), id(node->id),
       current_weight_vector_index(node->selected_weight_vector_index()) { 
     assert(!node->WeightVectors().empty());
     for (auto& wv : node->WeightVectors()) {
@@ -29,7 +29,7 @@ class GainBucketEntry {
     return all_weight_vectors[current_weight_vector_index];
   };
 
-  int gain;
+  double gain;
   int id;
   size_t current_weight_vector_index;
   std::vector<std::vector<int>> all_weight_vectors;
