+++ Graphs/gain_bucket_manager_multi_resource_mixed.cpp
@@ -432,7 +432,7 @@ bool GainBucketManagerMultiResourceMixed::Empty() const {
   return NumUnlockedNodes() == 0;
 }
 
-void GainBucketManagerMultiResourceMixed::AddNode(int gain, Node* node,
+void GainBucketManagerMultiResourceMixed::AddNode(double gain, Node* node,
     bool in_part_a, const std::vector<int>& total_weight) {
   GainBucketEntry entry(gain, node);
   if (in_part_a) {
@@ -502,7 +502,7 @@ int GainBucketManagerMultiResourceMixed::DetermineResourceAffinity(
 }
 
 void GainBucketManagerMultiResourceMixed::UpdateGains(
-    int gain_modifier, const vector<int>& nodes_to_increase_gain,
+    double gain_modifier, const vector<int>& nodes_to_increase_gain,
     const vector<int>& nodes_to_decrease_gain, bool moved_from_part_a) {
   vector<vector<int>> inc;  
   inc.resize(num_resources_per_node_);
